Better rating visiting
Part of #671
